% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_internal.R
\name{.AddHTTPArgs}
\alias{.AddHTTPArgs}
\title{Add GET parameters to a URL string.}
\usage{
.AddHTTPArgs(url, http_args)
}
\arguments{
\item{url}{A character vector containing the portion of the FIRST API URL
that follows the season parameter and forward slash, but is not a GET
parameter (i.e., before the '?'). For most FIRST API commands, the url
string will be the simple string that identifies the command, e.g.,
"districts" or "teams". A few FIRST API commands use path parameters (i.e.,
Detailed Scores) in addition to GET parameters -- in such cases the url
argument must contain the path parameters separated by '/'.}

\item{http_args}{A list of all GET parameters that will be supplied to the
URL query string. The elements of the list must have a name that that
matches the name of the GET parameter (e.g., teamNumber, eventCode). Any
list elements that are NULL will be skipped.}
}
\value{
A character vector containing the portion of the FIRST API URL
  starting with the path parameter that immediately follows the season (not
  including the '/') and extending to the end of the URL, including the GET
  query string.
}
\description{
.AddHTTPArgs is an internal function. It takes a named list and creates a GET
parameter string suitable for inclusion in a URL. It adds a question mark
('?') to the beginning of the string and separates each name-value pair with
an ampersand ('&'). List items with NULL values are ignored.
}
\examples{
  # From GetTeams()
  team_args <- list(teamNumber = team, eventCode = event,
    districtCode = district, state = state, page = page)
 url <- .AddHTTPArgs("teams", team_args)

 # From GetRankings()
 rank_args <- list(teamNumber = team, top = top)
 url <- .AddHTTPArgs(paste("rankings", event, sep = "/"), rank_args)
}
