% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_internal.R
\name{.FactorColumns}
\alias{.FactorColumns}
\title{Convert character columns to factors}
\usage{
.FactorColumns(df, cols)
}
\arguments{
\item{df}{The data.frame that will have it's columns converted to factors.}

\item{cols}{A character vector of column names that identifies the columns
that will be converted to factors.}
}
\value{
Data.frame
}
\description{
\code{.FactorColumns} is an internal function that is not intended to be
called by the user. It is called by other FIRST API methods.
}
\details{
There are benefits to converting data frame columns with a limited number of
unique character values to factors. For example, the R barplot function will
automatically group values by factors, which could be a team, district, etc.
Factors are conceptually similar to enumerations in other programming
languages.
}
\examples{
teams <- .FactorColumns(teams, c("districtCode", "stateProv", "country"))
}
