% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_internal.R
\name{.GetLocalData}
\alias{.GetLocalData}
\title{Extract FIRST data from a local file.}
\usage{
.GetLocalData(url, data_format)
}
\arguments{
\item{url}{A character vector containing the portion of the FIRST API url
that follows the season path parameter, not including the '/' that follows
the season.}

\item{data_format}{A character vector containing either "json" or "xml".}
}
\value{
While the local data will be of the same type as what was requested
  (i.e., if called via \code{GetTeams} a list of teams will be returned,
  \code{GetSchedule}) will return schedule data, etc.) arguments such as
  \code{team} or \code{event} will be ignored. Refer to the
  \code{data-raw/data.R} file to see the firstapiR commands that were used to
  download the data from the FIRST API server.
}
\description{
\code{.GetLocalData} is an internal function that is not intended to be
called by the user. It is called by other FIRST API methods.
}
\details{
\code{.GetLocalData} extracts archived FIRST data from the packages local
\code{R/sysdata.rda} file. The data is archived as JSON text or as an
XML::XMLDocument object.

Normally, \code{.GetLocalData} will not be used. However, if the user sets
the authorization key in the Session object to the literal value "key", then
\code{.GetHTTP()} will call \code{.GetLocalData()} instead of sending the
HTTP request. As a result, \code{.GetHTTP} and all other firstapiR functions
that depend on \code{.GetHTTP} will skip the HTTP request and return the
archived, local data.

This feature allows users to experiment with firstapiR features even if they
don't have an official authorization key issued by FIRST. It also enables
the examples in the firstapiR documentation to work. The actual HTTP request,
HTTP error code check, and extraction of text data from the body of the HTTP
response are the only parts of the code that are bypassed.
}
\examples{
content <- .GetLocalData(url, session$format)
}
