% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_internal.R
\name{.PreserveAttributes}
\alias{.PreserveAttributes}
\title{Reassign firstapiR attributes to a data frame.}
\usage{
.PreserveAttributes(df, att)
}
\arguments{
\item{df}{A data frame that may have had its attributes removed.}

\item{att}{A list containing the attributes from the data frame before the
attributes were stripped. This character vector should be created by
calling \code{attribues} on the original data frame.}
}
\value{
A data frame with the attributes \emph{local_test_data},
\emph{time_downloaded}, and \emph{last_modified}.
}
\description{
This is an internal function. Many actions in R will strip attributes from
data frames. This helper function replaces those attributes.
}
\details{
This function throws an error if the list passed via the \emph{att} argument
does not contain elements named \emph{local_test_data},
\emph{time_downloaded}, or \emph{last_modified}.
}
\examples{
new.df <- .PreserveAttributes(old.df)
}
