% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_internal.R
\name{.SetColumnOrder}
\alias{.SetColumnOrder}
\title{Set the column order of a data frame.}
\usage{
.SetColumnOrder(df, col.order)
}
\arguments{
\item{df}{A data frame}

\item{col.order}{A character vector containing the names of the columns in
\code{df}, in the desired column order.}
}
\value{
A data frame identical to the the data frame provided in the
  \code{df} argument, except with the columns in the order specified in the
  \code{col.order} argument. The returned data frame will ahve the same
  attributes as the data frame provided in the \code{df} argument.
}
\description{
\code{.SetColumnOrder()} is an internal function that is not intended to be
called by the user. It is called by other FIRST API methods.
}
\examples{
my.df <- data.frame(col3 = 1:3, col1 = 4:6, col2: 7:9)
ordered.df <- .SetColumnOrder(my.df, c("col1", "col2", "col3"))
}
