% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_internal.R
\name{.TrimColNames}
\alias{.TrimColNames}
\title{Remove prefixes from data frame column names}
\usage{
.TrimColNames(col_names)
}
\arguments{
\item{col_names}{A character vector of column names, generally provided by
the names() function.}
}
\value{
A character vector of trimmed column names, with the same length as
  the col_names argument.
}
\description{
\code{.TrimColNames} is an internal function that is not intended to be
called by the user. It is called by other FIRST API methods.
}
\details{
.TrimColNames removes all portions of a string up to and including the first
period. It's intended to produce shorter column names that are easier to
type in R interactive sessions. Column names without periods are left
unchanged.
}
\examples{
names(rankings) <- .TrimColNames(names(rankings))
}
