% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR.R
\docType{data}
\name{FRC_Data}
\alias{FRC_Data}
\alias{archimedes2016}
\alias{carson2016}
\alias{carver2016}
\alias{curie2016}
\alias{einstein2016}
\alias{galileo2016}
\alias{hopper2016}
\alias{newton2016}
\alias{tesla2016}
\title{2016 FRC Championship Data}
\format{An object of class \code{list} of length 16.}
\usage{
archimedes2016

carson2016

carver2016

curie2016

galileo2016

hopper2016

newton2016

tesla2016

einstein2016
}
\description{
firstapiR includes datasets for the 2016 FRC Championship. Each data set was
produced by the \code{firstapiR::GetAll()}and consists of a list of data
frames that were returned by the various event-related Get functions. See the
documentation for \link{\code{GetAll()}} for details on the contents of these
datasets.
}
\details{
There is a data set for every championship division, and for the final
championship matches on the Einstein field. The einstein2016 dataset is
smaller than the other datasets because it doesn't have qualification or
alliance selection data. Each dataset is stored in it's own rda file in the
package's data directory.
}
\seealso{
\link{\code{GetAll()}}
}
\keyword{datasets}

