% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_util.R
\name{GetAll}
\alias{GetAll}
\title{Download all data for a specific FRC competition.}
\usage{
GetAll(session, event)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}
}
\value{
A list containing the following firstapiR elements which are data
  frames, XML character vectors, or JSON character vectors:
  \enumerate{
    \item \emph{season}: Season class. See \link{GetSeason}.
    \item \emph{event}: Events class. See \link{GetEvents}.
    \item \emph{teams}: Teams class, listing all teams in the competition.
      See \link{GetTeams}.
    \item \emph{matches_qual}: MatchResults class, for qualification matches.
      See \link{GetMatchResults}.
    \item \emph{matches_playoff}: MatchResults class, for playoffs. See
      \link{GetMatchResults}.
    \item \emph{schedule_qual}: Schedule class, for qualification matches.
      See \link{GetSchedule}.
    \item \emph{schedule_playoff}: Schedule class, for playoffs. See
      \link{GetSchedule}.
    \item \emph{scores_qual}: Scores class, for qualification matches. See
      \link{GetScores}.
    \item \emph{scores_playoff}: Scores class, for playoffs. See
      \link{GetScores}.
    \item \emph{rankings}: Rankings class. See \link{GetRankings}.
    \item \emph{alliances}: Alliances class. See \link{GetAlliances}.
    \item \emph{awards}: Awards class. See \link{GetAwards}.}
\code{GetAll} will omit qualification-level elements and rankings if the
\code{event} argument is set to "CMP", which corresponds to the finals on
Einstein.
}
\description{
Downloads all data for a specific FRC competition and returns the data in a
list of data frames, or character vectors containing XML or JSON text,
depending on the format setting in the session variable. The list's named
elements are listed below in the \emph{Value} section.
}
\examples{
\dontrun{
sn <- GetSession("myUserName", "key", season = 2016)
archimedes <- GetAll(sn, "ARCHIMEDES")
SaveData(archimedes)
}
}
\seealso{
\link{GetEvents}
}

