% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetAlliances}
\alias{GetAlliances}
\title{Get playoff alliances}
\usage{
GetAlliances(session, event)
}
\arguments{
\item{session}{Session A session list created with \code{GetSession()}.}

\item{event}{Character Case insensitive event code (see \code{GetEvents()}).}
}
\value{
A data.frame, json list, or xml list.
   data.frame column names and classes:
     number: integer
     name: character
     captain, round1, round2, round3: integer
     backup, backupReplaced: integer
     count: integer
   Attribute "FIRST_type": "Alliances"
   Attribute "url": URL submitted to FIRST API
}
\description{
Returns a list of playoff alliances, including the alliance captains and
teams that were selected for the each alliance.
}
\details{
See the \emph{Event Alliances} section of the FIRST API documentation. The
URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/alliances/event}
}
\examples{
sn <- GetSession(username, key)
GetAlliances(sn, 'WAAMV')
}

