% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetAwards}
\alias{GetAwards}
\title{Get the awards that were presented to a team or at an event.}
\usage{
GetAwards(session, event = NULL, team = NULL, ...)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}). Must specify the \code{event} argument, the
\code{team} argument, or both.}

\item{team}{An integer vector containing a team number. Optional.  Must
specify the \code{event} argument, the \code{team} argument, or both.}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, a logical vector of length one with value
  \code{NA}, or a data.frame with class set to c("data.frame, "Awards").

  \strong{Data Frame Columns}
  \enumerate{
     \item \emph{awardId}: integer
     \item \emph{teamId}: integer
     \item \emph{eventId}: integer
     \item \emph{eventDivisionId}: logical
     \item \emph{eventCode}: character
     \item \emph{name}: character
     \item \emph{series}: integer
     \item \emph{teamNumber}: integer
     \item \emph{fullTeamName}: character
     \item \emph{person}: character}

  \strong{Data Frame or NA Object Attributes}
  \itemize{
    \item \emph{url}: Character vector containing URL submitted to FIRST API
      server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.
    \item \emph{local_url}: Character vector containing URL used to download
      local data.
    \item \emph{time_downloaded}: Character vector containing the local
      system time that the object was downladed from the FIRST API server.
      Formatted an an http date and time string.
    \item \emph{last_modified}: Character vector containing the date and time
      that the data was last modified by the FIRST API server.
    \item \emph{mod_since}: Character vector containing the value passed to
      the \code{mod_since argument}, or NULL if no argument was passed.
      Formatted as an http date and time string.
    \item \emph{only_mod_since}: Character vector containing the value passed
      to the \code{only_mod_since} argument, or NULL if no argument was
      passed. Formatted as an http date and time string.}
}
\description{
See the \emph{Event Awards} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.
}
\details{
The FIRST API URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/awards/event/team}
}
\examples{
sn <- GetSession("username", "key")
GetAwards(sn, "PNCMP")
GetAwards(sn, team = 360)
GetAwards(sn, "PNCMP", 360)
}

