% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetAwardsList}
\alias{GetAwardsList}
\title{Get a list of all available awards for a season.}
\usage{
GetAwardsList(session)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to
  c("data.frame", "Schedule").

  \strong{Data Frame Columns}
  \enumerate{
     \item \emph{awardId}: integer
     \item \emph{eventType}: character
     \item \emph{description}: character
     \item \emph{forPerson}: logical}

  \strong{Data Frame Attributes}
    \itemize{
    \item \emph{url}: URL submitted to FIRST API
    \item \emph{time_downloaded}: Local System time that the object was
    downladed from the FIRST API server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.}
}
\description{
See the \emph{Awards Listing} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.
}
\details{
The URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/awards/list}
}
\examples{
sn <- GetSession("username", "key")
GetAwardsList(sn)
}

