% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapi_http.R
\name{GetMatchResults}
\alias{GetMatchResults}
\title{Get Match results}
\usage{
GetMatchResults(session, event, level = "qual", team = NULL, match = NULL,
  start = NULL, end = NULL, expand_cols = FALSE)
}
\arguments{
\item{session}{Session A session list created with \code{GetSession()}.}

\item{event}{Character A FIRST API event code (see \code{GetEvents()}).}

\item{level}{Character Either \code{qual} or \code{playoff}. Defaults to
\code{qual}.}

\item{team}{Integer Team number. Optional. Returns only results for
matches in wich the team participated. Cannot specify \code{match} when
\code{team} is specified.}

\item{match}{Integer Returns results for only the specified match. Must
specify level when \code{match} is specified. Cannot specify \code{team}
when \code{match} is specified.}

\item{start}{Integer Eearliest match to return. Optional. Must specify level
when \code{start} is specified.}

\item{end}{Integer Latest match to return. Optional. Must specify level when
\code{end} is specified.}

\item{expand_cols}{A logical value. Optional, defaults to FALSE. If TRUE, the
dataframe will include one row for each completed match, with a different
column for each team. If FALSE, there will be six rows for each match, with
each row listing one assigned team and their station.}
}
\value{
A data.frame, json list, or xml list.
   data.frame column names and classes:
     actualStartTime: character
     description: character
     tournamentLevel: factor
     matchNumber: integer
     postResultTime: character

     If expand_rows == FALSE
       scoreRedFinal, scoreRedAuto, scoreRedFoul: integer
       scoreBlueFinal, scoreBlueAuto, scoreBlueFoul: integer
       Red1.team, Red2.team, Red3.team: factor
       Blue1.team, Blue2.team, Blue3.team: factor
       Red1.surrogate, Red2.surrogate, Red3.surrogate: logical
       Blue1.surrogate, Blue2.surrogate, Blue3.surrogate: logical
     If expand.rows == TRUE
       teamNumber: factor
       station: factor (Red1, Red2, Red3, Blue1, Blue2, Blue3)
       surrogate: logical
       scoreFinal, scoreAuto, scoreFoul: integer
     Attribute "FIRST_type": "MatchResults"
     Attribute "url": URL submitted to FIRST API
}
\description{
See the \emph{Match Results} section of the FIRST API documentation. The
URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/matches/event
?tournamentLevel=level&teamNumber=team&matchNumber=match&start=start&end=end}
}
\examples{
sn <- GetSession(username, key)
GetMatchResults(sn, 'PNCMP', level='qual')
GetMatchResults(sn, 'PNCMP', team='2990')
GetMatchResults(sn, 'WAAMV', match=2, level='playoff')
GetMatchResults(sn, 'CMP-ARCHIMEDES', level='qual', start=10, end=20)
}

