% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetRankings}
\alias{GetRankings}
\title{Get team rankings}
\usage{
GetRankings(session, event, team = NULL, top = NULL, ...)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}

\item{team}{An integer vector containing a team number. Optional.}

\item{top}{An integer vector specifying the number of teams to return,
starting with the top ranked team.}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, a logical vector of length one with value
  \code{NA}, or a data.frame with class set to c("data.frame, "Rankings").

  \strong{Data Frame Columns}
  \enumerate{
     \item \emph{rank}: integer
     \item \emph{teamNumber}: integer
     \item \emph{sortOrder1, sortOrder2, sortOrder3, sortOrder4, sortOrder5,
       sortOrder6}: integer or numeric.
     \item \emph{wins, losses, ties}: integer
     \item \emph{qualAverage}: numeric
     \item \emph{dq}: integer
     \item \emph{matchesPlayed}: integer}

  \strong{Data Frame or NA Object Attributes}
  \itemize{
    \item \emph{url}: Character vector containing URL submitted to FIRST API
      server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.
    \item \emph{local_url}: Character vector containing URL used to download
      local data.
    \item \emph{time_downloaded}: Character vector containing the local
      system time that the object was downladed from the FIRST API server.
      Formatted an an http date and time string.
    \item \emph{last_modified}: Character vector containing the date and time
      that the data was last modified by the FIRST API server.
    \item \emph{mod_since}: Character vector containing the value passed to
      the \code{mod_since argument}, or NULL if no argument was passed.
      Formatted as an http date and time string.
    \item \emph{only_mod_since}: Character vector containing the value passed
      to the \code{only_mod_since} argument, or NULL if no argument was
      passed. Formatted as an http date and time string.}
}
\description{
The results vary depending on the season requested. The 2016 data fields are
listed here. See the FIRST API documentation for data fields for prior
seasons.
}
\details{
See the \emph{Event Rankings} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/rankings/event?teamNumber=team&top=top}
}
\examples{
sn <- GetSession("username", "key")
GetRankings(sn, "WAAMV")
GetRankings(sn, "PNCMP", team = 1983)
GetRankings(sn, "ARCHIMEDES", top = 5)
}

