% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapi_http.R
\name{GetScores}
\alias{GetScores}
\title{Get detailed match scores.}
\usage{
GetScores(session, event, level = "qual", team = NULL, match = NULL,
  start = NULL, end = NULL)
}
\arguments{
\item{session}{Session A session list created with \code{GetSession()}.}

\item{event}{Character A FIRST API event code (see \code{GetEvents()}).}

\item{level}{Character Either \code{qual} or \code{playoff}. Defaults to
\code{qual}.}

\item{team}{Integer Team number. Optional. Returns only results for
matches in wich the team participated. Cannot specify \code{match} when
\code{team} is specified.}

\item{match}{Integer Returns results for only the specified match. Must
specify level when \code{match} is specified. Cannot specify \code{team}
when \code{match} is specified.}

\item{start}{Integer Eearliest match to return. Optional. Must specify level
when \code{start} is specified.}

\item{end}{Integer Latest match to return. Optional. Must specify level when
\code{end} is specified.}
}
\value{
A data.frame, json list, or xml list.
   data.frame column names and classes (2016):
     matchLevel: character
     matchNumber: integer
     audienceGroup: character
     alliance: character
     robot1Auto, robot2Auto, robot3Auto: character
     autoBouldersLow, autoBouldersHigh: integer
     teleopBouldersLow, teleopBouldersHigh: integer
     towerFaceA, towerFaceB, towerFaceC: character
     towerEndStrength: integer
     teleopTowerCaptured, teleopDefensesBreached: logical
     position1, position2, position3, position4, position5: character
     position1Crossings, position2Crossings, position3Crossings,
       position4Crossings, position5Crossings: integer
     foulCount, techFoulCount: integer
     autoPoints, autoReachPoints, autoCrossingPoints,
       autoBoulderPoints: integer
     teleopPoints, teleopCrossingPoints, teleopBoulderPoints,
       teleopChallengePoints, teleopScalePoints: integer
     breachPoints, capturePoints: integer
     adustPoints, foulPoints, totalPoints: integer
   Attribute "FIRST_type": "Scores"
   Attribute "url": URL submitted to FIRST API
}
\description{
The results vary depending on the season requested. The 2016 data fields are
listed here. See the FIRST API documentation for data fields for prior
seasons. The data frame contains two rows for each match, one for blue
and the other for red.
}
\details{
See the \emph{Detailed Scores} section of the FIRST API documentation. The
URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/matches/event/level?
teamNumber=team&matchNumber=match&start=start&end=end}
}
\examples{
sn <- GetSession(username, key)
GetScores(sn, event = "ARCHIMEDES")
GetScores(sn, event = "WAELL", start = 1, end = 10)
GetScores(sn, event = "WAELL", match = 15)
}

