% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetServerStatus}
\alias{GetServerStatus}
\title{Get the status of the FIRST API server.}
\usage{
GetServerStatus(session)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, a logical vector of length one with value
  \code{NA}, or a data.frame with class set to c("data.frame, "Status").

  \strong{Data Frame Columns}
  \enumerate{
    \item \emph{name}: character
    \item \emph{version}: double
    \item \emph{status}: character}

  \strong{Data Frame or NA Object Attributes}
  \itemize{
    \item \emph{url}: Character vector containing URL submitted to FIRST API
      server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.
    \item \emph{local_url}: Character vector containing URL used to download
      local data.
    \item \emph{time_downloaded}: Character vector containing the local
      system time that the object was downladed from the FIRST API server.
      Formatted an an http date and time string.
    \item \emph{last_modified}: Character vector containing the date and time
      that the data was last modified by the FIRST API server.}
}
\description{
See the \emph{API Index} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.
}
\details{
The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0}
}
\examples{
sn <- GetSession("username", "key")
status <- GetServerStatus(sn)
}

