% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapi_http.R
\name{GetSession}
\alias{GetSession}
\title{Create a FIRST API session.}
\usage{
GetSession(username, key, season = .default_season, format = "data.frame",
  staging = FALSE)
}
\arguments{
\item{username}{Character The username assigned by FIRST.}

\item{key}{Character The authorization key assigned by FIRST.}

\item{season}{Integer The 4-digit year. Defaults to the year specified in the
.default_season variable at the top if this file. Must be a 4-digit number
  that is equal to or less than the current year and greater than or equal to
  the current year plus one.}

\item{format}{Character The data format that will be returned by the
functions. Can be "json", "data.frame", or "xml". Defaults to "data.frame".
Case insensitive.}

\item{staging}{Logical Set to TRUE to use the staging URL. Defaults to
  \code{FALSE}.

Throws an error if season, format, or staging arguments are not allowed
values.}
}
\value{
A list containing all GetSession parameters. The list has an
  attribute, "FIRST_type", that is set to "Session".
}
\description{
Every FIRST API function requires a session as its first parameter.
}
\details{
The session is an R list that contains the FIRST API username, authorization
key, season, format, and a boolean value that specifies whether to use
the staging server instead of the production server.
}
\examples{
sn <- GetSession('myUserName', 'myAuthorizationKey')
sn <- GetSession('myUserName', 'myAuthorizationKey', season = 2015)
sn$format <- 'xml'
}

