% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapi_http.R
\name{GetTeams}
\alias{GetTeams}
\title{Get details for many teams.}
\usage{
GetTeams(session, team = NULL, event = NULL, district = NULL,
  state = NULL, page = NULL)
}
\arguments{
\item{session}{Session A session list created with \code{GetSession()}.}

\item{team}{Integer Team number. Optional}

\item{event}{Character A FIRST API event code (see \code{GetEvents()}). If
event is specified, \code{GetTeams()} will filter results to all teams
particpating in the specified event. Optional.}

\item{district}{Character The FIRST API district code (see
\code{GetDistricts()}). If disctrict is specified, \code{GetTeams()} will
filter results to all teams in the specified district.}

\item{state}{Character A state name, spelled out entirely (i.e., 'Idaho',
\emph{not} 'ID'). If state is specified, \code{GetTeams()} will filter
results to all teams in the specified state.}

\item{page}{Integer Page number specifying which page of results should be
returned. Optional. Use only for xml or json formats.}
}
\value{
If the data.frame format is specified (i.e., \code{session$format ==
  'data.frame'}), returns all teams in a single data frame. If the json or
  xml formates are specified, returns a single page of json or xml responses.
   data.frame column names and classes:
     teamNumber: character
     nameFull: character
     nameShort: character
     city: character
     stateProv: factor
     country: factor
     website: character
     rookieYear: integer
     robotName: character
     districtCode: factor
     teamCountTotal: integer
     teamCountPage: integer
     pageCurrent: integer
     pageTotal: integer
   Attribute "FIRST_type": "Teams"
   Attribute "url": URL submitted to FIRST API
}
\description{
The FIRST team listing API response will be broken up into several pages if
the number of teams in the response exceeds 65, with a separate HTTP request
required for each page. If the dataframe format is specified, then
\code{GetTeams()} makes this transparent to the user. \code{GetTeams()} will
determine the number of pages in the response, conduct an HTTP request to
obtain each page, and merge all pages into a single dataframe. For the xml
and json formats, \code{GetTeams()} will return a single page of teams.
}
\details{
See the \emph{Team Listings} section of the FIRST API documentation. The
URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/teams&eventCode=event
?districtCode=district?state=state?page=2}
}
\examples{
sn <- GetSession(username, key)
GetTeams(state = 'California')
GetTeams(district = 'FIM')
GetTeams(event = 'CMP-CARVER')
}

