% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapi_analyze.R
\name{Points_boxplot}
\alias{Points_boxplot}
\title{Analysis Module, Version 1.0}
\usage{
Points_boxplot(match_results, rankings, event, start = 1, stop = 10)
}
\description{
Analysis Module, Version 1.0
}

