% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_util.R
\name{ReadData}
\alias{ReadData}
\title{Reads data from an RDS data file via a file open dialog.}
\usage{
ReadData()
}
\value{
The R object stored in the RDS data file, or NA if the user cancels
  the dialog without selecting a file.
}
\description{
Displays the tcltk package's \emph{save file dialog}. Once the user enters or
selects a file name, \code{ReadData()} will read and return the value stored
in the corresponding RDS data file. The dialog will initially display
whatever folder the current working directory (see \code{getwd()} and
\code{setwd()}).
}
\details{
ReadData uses the \code{readRDS()} function, from R's base package, to read
R data. The data is stored in the RDS file as an unnamed object, so the user
must assign the return value to a named variable.

Throws an error if the selected file does not end in "RDS"
(case-insensitive). Prints a message if the user cancels the dialog without
selecting a file.
}
\examples{
\dontrun{
my.data <- ReadData()
}
}

