% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR.R
\name{Standard_attributes}
\alias{Standard_attributes}
\title{Standard firstapiR Attributes}
\description{
Objects returned by firstapiR functions that begin with \emph{Get} have the
following attributes:
}
\details{
\itemize{
  \item \emph{url}: Character vector containing URL submitted to FIRST API
    server.
  \item \emph{local_test_data}: \code{TRUE} if data was extracted from
    R/sysdata.rda file.
  \item \emph{local_url}: Character vector containing URL used to download
    local data.
  \item \emph{time_downloaded}: Character vector containing the local
    system time that the object was downladed from the FIRST API server.
    Formatted an an http date and time string.
  \item \emph{last_modified}: Character vector containing the date and time
    that the data was last modified by the FIRST API server.
  \item \emph{mod_since}: Character vector containing the value passed to
    the \code{mod_since argument}, or NULL if no argument was passed.
    Formatted as an http date and time string.
  \item \emph{only_mod_since}: Character vector containing the value passed
    to the \code{only_mod_since} argument, or NULL if no argument was
    passed. Formatted as an http date and time string.}

Use the \code{attr()} function to retrieve the value of these attributes.
}

