% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_util.R
\name{ToMatchShape}
\alias{ToMatchShape}
\title{Convert a Schedule, Hybrid, or MatchResults data frame to match shape.}
\usage{
ToMatchShape(df)
}
\arguments{
\item{username}{A firstapiR data frame. The data frame must inherit from
the schedule, hybridSchedule, or matchResults class data frames returned
by firstapiR functions. The data frame must also be in team shape and have
the shape attribute set to "team".}
}
\value{
A data frame of class schedule, hybridSchedule, or matchResults
  (depending on the class of the df argument) in match shape, with the match
  attribute set to "match".
}
\description{
The firstapiR functions \code{GetSchedule()}, \code{GetHybridSchedule()}, and
\code{GetMatchResults} return a data frame in team shape, i.e., with one team
listed in every row, and six rows per match. R documentation often refers to
data frames in this format as \emph{narrow} data frames. \code{ToMatchShape}
will convert these narrow data frames to a wide data frame (it's wide because
it has more columns) with all six teams listed in one row and only one row
per match.
}
\details{
ToMatchShape will only convert data frames that are in team shape, with their
shape attribute set to "team". To convert data frame that are in alliance
shape to match shape, first use \code{ToTeamShape} to convert the data frame
back to team shape, and then use \code{ToMatchShape} to convert it to match
shape.

FirstapiR proivdes functions to reshape data frames because different shapes
are useful for different purposes. For example, team shape is useful for
extracting data for a single team because it's only necessary to subset the
data frame on the single team column. Match shape is useful for displaying
schedules. Alliance shape is useful for calculating offensive power rating
(OPR).

This function throws an error if df does not inherit from
\emph{matchResults}, \emph{schedule}, or \emph{hybridSchedule}, or if the
data frame's shape attribute is missing or is anything other than
\emph{team}.
}
\examples{
sn <- GetSession("myUserName", "key")
matchResults <- GetMatchResults(sn, event = "PNCMP")
matchResults.m <- ToMatchShape(matchResults)
}

