% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_util.R
\name{ToTeamShape}
\alias{ToTeamShape}
\title{Convert a Schedule, Hybrid, or MatchResults data frame to team shape.}
\usage{
ToTeamShape(df)
}
\arguments{
\item{df}{A firstapiR data frame. The data frame must inherit from the
\emph{schedule}, \emph{hybridSchedule}, or \emph{matchResults} class data
frames returned by firstapiR functions. The data frame must also be in team
shape and have the shape attribute set to "match" or "alliance".}
}
\value{
A data frame of class \emph{schedule}, \emph{hybridSchedule}, or
  \emph{matchResults} (depending on the class of the df argument) in match
  shape, with the shape attribute set to "match".
}
\description{
The firstapiR functions \code{GetSchedule()}, \code{GetHybridSchedule()}, and
\code{GetMatchResults} return a data frame in team shape, i.e., with one team
listed in every row, and six rows per match. R documentation often refers to
data frames in this format as \emph{narrow} data frames. The functions
\code{ToMatchShape} and \code{ToAllianceShape} will convert data frames in
the narrow team shape to the wide (they're called wide because they have more
columns) alliance or match shapes. \code{ToTeamShape} converts these wide
data frames back to narrow data frames in team shape.
}
\details{
\code{ToTeamShape} will only convert data frames that are in match or
alliance shape, with their shape attribute set to \emph{match} or
\emph{alliance}.

FirstapiR proivdes functions to reshape data frames because different shapes
are useful for different purposes. For example, team shape is useful for
extracting data for a single team because it's only necessary to subset the
data frame on the team column. Match shape is useful for displaying
schedules. Alliance shape is useful for calculating offensive power rating
(OPR).

This function throws an error if df does not inherit from
\emph{matchResults}, \emph{schedule}, or \emph{hybridSchedule}, or if the
data frame's shape attribute is missing or is anything other than
\emph{team}.
}
\examples{
sn <- GetSession("myUserName", "key", season = 2016)
matchResults_teamshape <- GetMatchResults(sn, event = "PNCMP")
matchResults_matchshape <- ToMatchShape(matchResults_teamshape)
matchResults_back_to_teamshape <- ToTeamShape(matchResults_matchshape)
}
\seealso{
\code{\link{GetMatchResults}}, \code{\link{GetHybridSchedule}},
  \code{\link{GetSchedule}}
}

