% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_util.R
\name{GetAll}
\alias{GetAll}
\title{Download all data for a competition, including schedules, scores, and awards.}
\usage{
GetAll(session, event)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}
}
\value{
A list containing the following firstapiR data frames:
  \enumerate{
    \item \emph{season}: Season class
    \item \emph{event}: Events class
    \item \emph{teams}: Teams class, listing all teams in the competition
    \item \emph{matches_qual}: MatchResults class, for qualification matches
    \item \emph{matches_playoff}: MatchResults class, for playoffs
    \item \emph{schedule_qual}: Schedule class, for qualification matches
    \item \emph{schedule_playoff}: Schedule class, for playoffs
    \item \emph{scores_qual}: Scores class, for qualification matches
    \item \emph{scores_playoff}: Scores class, for playoffs
    \item \emph{rankings}: Rankings class
    \item \emph{alliances}: Alliances class
    \item \emph{awards}: Awards class}
}
\description{
Download all data for a competition, including schedules, scores, and awards.
}
\examples{
\dontrun{
sn <- GetSession("myUserName", "key")
archimedes <- GetAll(sn, "ARCHIMEDES")
SaveData(ARCHIMEDES)
}
}

