% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_main.R
\name{GetAlliances}
\alias{GetAlliances}
\title{Get teams assigned to playoff alliances}
\usage{
GetAlliances(session, event, mod_since = NULL, only_mod_since = NULL)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to c("data.frame,
  "Alliances"). Returns a logical vector of length one with value \code{NA}
  if data is unchanged since date and time passed in arguments
  \code{mod_since} or \code{only_mod_since}.
}
\description{
Returns a list of playoff alliances, including the alliance captains and
teams that were selected for each alliance.
}
\details{
See the \emph{Event Alliances} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/alliances/event}
}
\section{Columns}{

  \enumerate{
     \item \emph{number}: integer
     \item \emph{name}: character
     \item \emph{captain, round1, round2, round3}: integer
     \item \emph{backup, backupReplaced}: integer
     \item \emph{count}: integer}
}
\examples{
sn <- GetSession("username", "key")
GetAlliances(sn, "WAAMV")
}
\seealso{
Refer to \code{\link{Standard firstapiR Attributes}} for data
attributes returned by this function.
}

