% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetAwards}
\alias{GetAwards}
\title{Get the awards that were presented to a team or at an event.}
\usage{
GetAwards(session, event = NULL, team = NULL)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}). Must specify the \code{event} argument, the
\code{team} argument, or both.}

\item{team}{An integer vector containing a team number. Optional.  Must
specify the \code{event} argument, the \code{team} argument, or both.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to
  c("data.frame", "Schedule").

  \strong{Data Frame Columns}
  \enumerate{
     \item \emph{awardId}: integer
     \item \emph{teamId}: integer
     \item \emph{eventId}: integer
     \item \emph{eventDivisionId}: logical
     \item \emph{eventCode}: character
     \item \emph{name}: character
     \item \emph{series}: integer
     \item \emph{teamNumber}: integer
     \item \emph{fullTeamName}: character
     \item \emph{person}: character}

  \strong{Data Frame Attributes}
    \itemize{
    \item \emph{url}: URL submitted to FIRST API
    \item \emph{time_downloaded}: Local System time that the object was
    downladed from the FIRST API server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.}
}
\description{
See the \emph{Event Awards} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.
}
\details{
The URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/awards/event/team}
}
\examples{
sn <- GetSession("username", "key")
GetAwards(sn, "PNCMP")
GetAwards(sn, team = 360)
GetAwards(sn, "PNCMP", 360)
}

