% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapi_http.R
\name{GetAwards}
\alias{GetAwards}
\title{Get the awards that were presented to a team or at an event.}
\usage{
GetAwards(session, event = NULL, team = NULL)
}
\arguments{
\item{session}{Session A session list created with \code{GetSession()}.}

\item{event}{Character Case insensitive event code (see \code{GetEvents()}).}

\item{team}{Integer Team number.}
}
\value{
A data.frame, json list, or xml list.
   data.frame column names and classes:
     awardId: integer
     teamId: integer
     eventId: integer
     eventDivisionId: logical
     eventCode: character
     name: character
     series: integer
     teamNumber: integer
     fullTeamName: character
     person: character
   Attribute "FIRST_type": "Awards"
   Attribute "url": URL submitted to FIRST API
}
\description{
See the \emph{Event Awards} section of the FIRST API documentation. The
URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/awards/event/team}
}
\examples{
sn <- GetSession(username, key)
GetAwards(sn, "PNCMP")
GetAwards(sn, team = 360)
GetAwards(sn, event = "PNCMP", 360)
}

