% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_main.R
\name{GetAwards}
\alias{GetAwards}
\title{Get the awards that were presented at an event}
\usage{
GetAwards(session, event = NULL, team = NULL, mod_since = NULL,
  only_mod_since = NULL)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}). Must specify the \code{event} argument, the
\code{team} argument, or both.}

\item{team}{An integer vector containing a team number. Optional.  Must
specify the \code{event} argument, the \code{team} argument, or both.}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object or a data.frame with class set to c("data.frame,
  "Awards"). Returns a logical vector of length one with value \code{NA} if
  data is unchanged since date and time passed in arguments \code{mod_since}
  or \code{only_mod_since}.
}
\description{
See the \emph{Event Awards} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.
}
\details{
The FIRST API URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/awards/event/team}
}
\section{Columns}{

  \enumerate{
     \item \emph{awardId}: integer
     \item \emph{teamId}: integer
     \item \emph{eventId}: integer
     \item \emph{eventDivisionId}: logical
     \item \emph{eventCode}: character
     \item \emph{name}: character
     \item \emph{series}: integer
     \item \emph{team}: integer
     \item \emph{fullTeamName}: character
     \item \emph{person}: character}
}
\examples{
sn <- GetSession("username", "key")
GetAwards(sn, "PNCMP")
GetAwards(sn, team = 360)
GetAwards(sn, "PNCMP", 360)
}
\seealso{
Refer to \code{\link{Standard_attributes}} for data
attributes returned by this function.
}

