% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetAwardsList}
\alias{GetAwardsList}
\title{Get a list of all available awards for a season.}
\usage{
GetAwardsList(session)
}
\arguments{
\item{session}{Session A list created with \code{GetSession()}.}
}
\value{
A data.frame, json list, or xml list.
   data.frame column names and classes:
     awardId: integer
     eventType: character
     description: character
     forPerson: logical
   Attribute "FIRST_type": "AwardsList"
   Attribute "url": URL submitted to FIRST API
}
\description{
See the \emph{Awards Listing} section of the FIRST API documentation. The
URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/awards/list}
}
\examples{
sn <- GetSession(username, key)
GetAwardsList(sn)
}

