% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetDistricts}
\alias{GetDistricts}
\title{Get a list of FIRST districts}
\usage{
GetDistricts(session)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to
  c("data.frame, "Districts").

  \strong{Data Frame Columns}
  \enumerate{
    \item \emph{code}: character
    \item \emph{name}: character
    \item \emph{districtCount}: integer}

  \strong{Data Frame Attributes}
  \itemize{
    \item \emph{url}: URL submitted to FIRST API
    \item \emph{time_downloaded}: Local System time that the object was downladed
      from the FIRST API server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.}
}
\description{
This function returns a list of all current districs, including their titles
and codes. District codes are used as parameters for several other FIRST API
functions.
}
\details{
See the \emph{District Listings} section of the FIRST API documentation.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/districts}
}
\examples{
sn <- GetSession(username, "key")
districts <- GetDistricts(sn)
}

