% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetDistricts}
\alias{GetDistricts}
\title{Get a list of FIRST districts.}
\usage{
GetDistricts(session)
}
\arguments{
\item{session}{Session A session list created with \code{GetSession()}.}
}
\value{
A data.frame, json list, or xml list.
   data.frame column names and classes:
     code: character
     name: character
     districtCount: integer
   Attribute "FIRST_type": "Districts"
   Attribute "url": URL submitted to FIRST API
}
\description{
This function returns a list of all current districs, including their titles
and codes. District codes are used as parameters for several other FIRST API
functions.
}
\details{
See the \emph{District Listings} section of the FIRST API documentation. The
URL format is: \code{https://frc-api.firstinspires.org/v2.0/season/districts}
}
\examples{
sn <- GetSession(username, key)
districts <- GetDistricts(sn)
}

