% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetEvents}
\alias{GetEvents}
\title{Get information about FRC events}
\usage{
GetEvents(session, event = NULL, team = NULL, district = NULL,
  exclude_district = NULL, ...)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code. If event
is specified, \code{GetEvents()} will return results only for the specified
event. Optional.}

\item{team}{An integer vector containing a team number. Optional}

\item{district}{A character vector containing the FIRST API district code
(see \code{GetDistricts()}). If \code{district} is specified,
\code{GetTeams()} will filter results to only the events in the specified
district.}

\item{exclude_district}{A logical vector. If set to \code{TRUE}, district
events are excluded from results. Optional.}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, a logical vector of length one with value
  \code{NA}, or a data.frame with class set to c("data.frame, "Events").

  \strong{Data Frame Columns}
  \enumerate{
    \item \emph{code}: character
    \item \emph{divisionCode}: character
    \item \emph{name}: character
    \item \emph{type}: factor ('Regional', 'DistrictEvent',
       'DistrictChampionship', 'ChampionshipSubdivision',
       'ChampionshipDivision', 'Championship', 'Offseason')
    \item \emph{districtCode}: factor ('CHM', 'FIM', 'IN', 'MAR', 'NC',
      'PCH', 'PNW')
    \item \emph{venue}: character
    \item \emph{city}: character
    \item \emph{stateprov}: factor
    \item \emph{country}: factor
    \item \emph{timezone}: factor
    \item \emph{dateStart}: character
    \item \emph{dateEnd}: character
    \item \emph{eventCount}: integer}

  \strong{Data Frame or NA Object Attributes}
  \itemize{
    \item \emph{url}: Character vector containing URL submitted to FIRST API
      server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.
    \item \emph{local_url}: Character vector containing URL used to download
      local data.
    \item \emph{time_downloaded}: Character vector containing the local
      system time that the object was downladed from the FIRST API server.
      Formatted an an http date and time string.
    \item \emph{last_modified}: Character vector containing the date and time
      that the data was last modified by the FIRST API server.
    \item \emph{mod_since}: Character vector containing the value passed to
      the \code{mod_since argument}, or NULL if no argument was passed.
      Formatted as an http date and time string.
    \item \emph{only_mod_since}: Character vector containing the value passed
      to the \code{only_mod_since} argument, or NULL if no argument was
      passed. Formatted as an http date and time string.}
}
\description{
See the \emph{Event Listings} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.
}
\details{
\code{GetEvents} will accept either the \code{team} or \code{district}
parameters, neither parameter, or both parameters. If neither \code{team} nor
\code{district} are specified, \code{GetEvents} returns all FRC events for
the competition season. If \code{team} is specified, the results are filtered
to only the events in which the FRC team participated. Similarly, if
\code{district} is specified, the results are filtered to only the events
that occurred within the specified district. If \code{exclude_district} is
set to TRUE, then only non-district events are returned. The \code{district}
and \code{exclude_district} events may not be specified at the same time.

Throws an error if \code{team} is specified and any other arguments are
specified, or if both the \code{district} and \code{exclude_district}
arguments are specified.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/events?
  teamNumber=team&districtCode=district&excludeDistrict=district}
}
\examples{
sn <- GetSession("username", "key")
frc_data <- GetEvents(sn, team = 5803)
frc_data <- GetEvents(sn, district = 'PNW')
frc_data <- GetEvents(sn, team = 360, exclude_district = TRUE)
}

