% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_main.R
\name{GetHybridSchedule}
\alias{GetHybridSchedule}
\title{Get the match schedule and results}
\usage{
GetHybridSchedule(session, event, level = "qual", start = NULL,
  end = NULL, mod_since = NULL, only_mod_since = NULL)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}

\item{level}{A character vector containing either \emph{"qual"} or
\emph{"playoff"}. Optional: defaults to \emph{"qual"}.}

\item{start}{An integer vector containing the earliest match to return.
Optional.}

\item{end}{An integer vector containing the latest match to return. Optional.}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to c("data.frame,
  "HybridSchedule"). Returns a logical vector of length one with value
  \code{NA} if data is unchanged since date and time passed in arguments
  \code{mod_since} or \code{only_mod_since}.
}
\description{
For matches that have been played, \code{GetHybridSchedule} returns the teams
assigned to the match and the match results. If the mtach has not yet been
played, the assigned teams and schedule data are returned, but the result
fields are blank.
}
\details{
The data frame returned by \code{GetHybridSchedule()} is in team shape,
i.e., each ' row contains data for a single team and there are six rows per
match. Use ' \code{ToAllianceShape()} or \code{ToMatchShape} to convert the
data frame to ' a three-teams-per-row shape or a six-teams-per-row shape.

See the \emph{Hybrid Schedule} section of the FIRST API documentation for
more details.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/schedule/event/level/
hybrid?start=start&end=end}
}
\section{Columns}{

  \enumerate{
     \item \emph{match}: integer
     \item \emph{description}: character
     \item \emph{level}: factor
     \item \emph{start}: character
     \item \emph{actualStart}: character
     \item \emph{team}: factor
     \item \emph{station}: factor (red1, red2, red3, blue1, blue2, blue3)
     \item \emph{surrogate}: logical
     \item \emph{disqualified}: logical
     \item \emph{scoreFinal, scoreAuto, scoreFoul}: integer}
}

\section{Attributes}{

  \itemize{
    \item \emph{url}: Character vector containing URL submitted to FIRST API
      server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.
    \item \emph{local_url}: Character vector containing URL used to download
      local data.
    \item \emph{shape}: Character vector that specifies the row and column
      arrangement of the data frame. Initially set to \emph{team},
      indicating there is one team per row and six rows per match. The
      functions \code{ToAllianceShape()} and \code{ToMatchShape} will reshape
      the data frame to have either three teams per row or six teams per row
      and will set the \emph{shape} attribute to \emph{alliance} or
      \emph{match}.
    \item \emph{time_downloaded}: Character vector containing the local
      system time that the object was downladed from the FIRST API server.
      Formatted an an http date and time string.
    \item \emph{last_modified}: Character vector containing the date and time
      that the data was last modified by the FIRST API server.
    \item \emph{mod_since}: Character vector containing the value passed to
      the \code{mod_since argument}, or NULL if no argument was passed.
      Formatted as an http date and time string.
    \item \emph{only_mod_since}: Character vector containing the value passed
      to the \code{only_mod_since} argument, or NULL if no argument was
      passed. Formatted as an http date and time string.}
}
\examples{
sn <- GetSession("username", "key")
GetHybridSchedule(sn, event = "ORPHI")
GetHybridSchedule(sn, event = "WAELL", level = "playoff", start = 3, end = 6)
}

