% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetHybridSchedule}
\alias{GetHybridSchedule}
\title{Get the match schedule and results.}
\usage{
GetHybridSchedule(session, event, level = "qual", start = NULL,
  end = NULL, expand_cols = FALSE, ...)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}

\item{level}{A character vector containing either \emph{"qual"} or
\emph{"playoff"}. Defaults to \emph{"qual"}. Optional.}

\item{start}{An integer vector containing the earliest match to
Optional.}

\item{end}{An integer vector containing the latest match to return. Optional.}

\item{expand_cols}{A logical value that defaults to \code{FALSE}. If
\code{TRUE}, the dataframe will include one row for each scheduled match,
with a different column for each team. If \code{FALSE}, there will be six
rows for each match, with each row listing one assigned team and their
station. Optional.}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, a logical vector of length one with value
  \code{NA}, or a data.frame with class set to c("data.frame,
  "HybridSchedule").

  \strong{Data Frame Columns}
  \enumerate{
     \item \emph{description}: character
     \item \emph{tournamentLevel}: factor
     \item \emph{matchNumber}: integer
     \item \emph{startTime}: character
     \item \emph{actualStartTime}: character}

     If expand_cols == \emph{FALSE}
       \enumerate{
         \item \emph{teamNumber}: factor
         \item \emph{station}: factor (Red1, Red2, Red3, Blue1, Blue2, Blue3)
         \item \emph{scoreFinal, scoreAuto, scoreFoul}: integer
         \item \emph{surrogate}: logical
         \item \emph{dq}: logical}

     If expand_cols == \emph{TRUE}
       \enumerate{
         \item \emph{scoreRedFoul, scoreRedAuto, scoreRedFinal}: integer
         \item \emph{scoreBlueFoul, scoreBlueAuto, scoreBlueFinal}: integer
         \item \emph{Red1.team, Red2.team, Red3.team}: factor
         \item \emph{Blue1.team, Blue2.team, Blue3.team}: factor
         \item \emph{Red1.surrogate, Red2.surrogate, Red3.surrogate}: logical
         \item \emph{Blue1.surrogate, Blue2.surrogate,
           Blue3.surrogate}: logical
         \item \emph{Red1.dq, Red2.dq, Red3.dq}: logical
         \item \emph{Blue1.dq, Blue2.dq, Blue3.dq}: logical}

  \strong{Data Frame or NA Object Attributes}
  \itemize{
    \item \emph{url}: Character vector containing URL submitted to FIRST API
      server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.
    \item \emph{local_url}: Character vector containing URL used to download
      local data.
    \item \emph{time_downloaded}: Character vector containing the local
      system time that the object was downladed from the FIRST API server.
      Formatted an an http date and time string.
    \item \emph{last_modified}: Character vector containing the date and time
      that the data was last modified by the FIRST API server.
    \item \emph{mod_since}: Character vector containing the value passed to
      the \code{mod_since argument}, or NULL if no argument was passed.
      Formatted as an http date and time string.
    \item \emph{only_mod_since}: Character vector containing the value passed
      to the \code{only_mod_since} argument, or NULL if no argument was
      passed. Formatted as an http date and time string.}
}
\description{
For matches that have been played, \code{GetHybridSchedule} returns the teams
assigned to the match and the match results. If the mtach has not yet been
played, the assigned teams and schedule data are returned, but the result
fields are blank.
}
\details{
Depending on the \code{expand_cols} argument, \code{GetHybridSchedule} can
structure the resulting data frame two diferent ways. When \code{expand_cols}
is set to \emph{FALSE} the data frame will contain six rows for every match
returned. Each row will include data on one team that participates in the
match. This narrow (i.e., fewer columns) structure is useful when filtering
results to specific teams, because only one column, \emph{teamNumber}, must
be filtered. When \code{expand_cols} is set to \emph{TRUE} the data frame
will have one row per match, with all six participating teams listed in one
row. This wide format is useful for displaying the schedule in a table.

See the \emph{Hybrid Schedule} section of the FIRST API documentation for
more details.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/schedule/event/level/
hybrid?start=start&end=end}
}
\examples{
sn <- GetSession("username", "key")
GetHybridSchedule(sn, event = "ORPHI")
GetHybridSchedule(sn, event = "WAELL", level = "playoff", start = 3, end = 6)
}

