% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_main.R
\name{GetMatchResults}
\alias{GetMatchResults}
\title{Get match scores and participating teams}
\usage{
GetMatchResults(
  session,
  event,
  level = "qual",
  team = NULL,
  match = NULL,
  start = NULL,
  end = NULL,
  mod_since = NULL,
  only_mod_since = NULL
)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}

\item{level}{A character vector containing either \emph{"qual"} or
\emph{"playoff"}. Defaults to \emph{"qual"}. Optional.}

\item{team}{An integer vector containing a team number. Optional. Cannot
specify \code{match} when \code{team} is specified.}

\item{match}{An integer vector containing a match number. Optional. If
specified, \code{GetMatchResults} returns results for only the specified
match. If \code{level} is not specified, returns the results for the
qualification match. To get playoff match results, set \code{level} to
\emph{"playoff"}. Cannot specify \code{team} when \code{match} is
specified.}

\item{start}{An integer vector containing the earliest match to return.
Optional.}

\item{end}{An integer vector containing the latest match to return. Optional.}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to c("data.frame,
  "MatchResults"). Returns a logical vector of length one with value
  \code{NA} if data is unchanged since date and time passed in arguments
  \code{mod_since} or \code{only_mod_since}.
}
\description{
The data frame returned by \code{GetMatchResults()} is in team shape, i.e.,
each row contains data for a single team and there are six rows per match.
Use \code{ToAllianceShape()} or \code{ToMatchShape()} to convert the data
frame to a three-teams-per-row shape or a six-teams-per-row shape.
}
\details{
See the \emph{Match Results} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/matches/event
?tournamentLevel=level&teamNumber=team&matchNumber=match&start=start&end=end}
}
\section{Columns}{

  \enumerate{
     \item \emph{match}: integer
     \item \emph{description}: character
     \item \emph{level}: factor
     \item \emph{actualStart}: character
     \item \emph{postResult}: character
     \item \emph{team}: factor
     \item \emph{alliance}: factor (Blue, Red)
     \item \emph{station}: factor (Red1, Red2, Red3, Blue1, Blue2, Blue3)
     \item \emph{disqualified}: logical
     \item \emph{scoreFinal, scoreAuto, scoreFoul}: integer}
}

\examples{
sn <- GetSession("username", "key", season = 2016)
GetMatchResults(sn, "PNCMP", level="qual")
GetMatchResults(sn, "PNCMP", team="2990")
GetMatchResults(sn, "WAAMV", match=2, level="playoff")
GetMatchResults(sn, "CMP-ARCHIMEDES", level="qual", start=10, end=20)
}
\seealso{
Refer to \code{\link{Standard_attributes}} for data
attributes returned by this function.
}
