% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_main.R
\name{GetRankings}
\alias{GetRankings}
\title{Get team rankings}
\usage{
GetRankings(session, event, team = NULL, top = NULL, mod_since = NULL,
  only_mod_since = NULL)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}

\item{team}{An integer vector containing a team number. Optional.}

\item{top}{An integer vector specifying the number of teams to return,
starting with the top ranked team.}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to c("data.frame,
  "Rankings"). Returns a logical vector of length one with value \code{NA} if
  data is unchanged since date and time passed in arguments \code{mod_since}
  or \code{only_mod_since}.
}
\description{
The results vary depending on the season requested. The 2016 data fields are
listed here. See the FIRST API documentation for data fields for prior
seasons.
}
\details{
See the \emph{Event Rankings} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/rankings/event?teamNumber=team&top=top}
}
\section{Columns}{

  \enumerate{
     \item \emph{rank}: integer
     \item \emph{team}: integer
     \item \emph{sortOrder1, sortOrder2, sortOrder3, sortOrder4, sortOrder5,
       sortOrder6}: integer or numeric.
     \item \emph{wins, losses, ties}: integer
     \item \emph{qualAverage}: numeric
     \item \emph{dq}: integer
     \item \emph{matchesPlayed}: integer}
}
\examples{
sn <- GetSession("username", "key")
GetRankings(sn, "WAAMV")
GetRankings(sn, "PNCMP", team = 1983)
GetRankings(sn, "ARCHIMEDES", top = 5)
}
\seealso{
Refer to \code{\link{standard_attributes}} for data
attributes returned by this function.
}

