% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetRankings}
\alias{GetRankings}
\title{Get team rankings}
\usage{
GetRankings(session, event, team = NULL, top = NULL)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}

\item{team}{An integer vector containing a team number. Optional.}

\item{top}{An integer vector specifying the number of teams to return,
starting with the top ranked team.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to
  c("data.frame", "Schedule").

  \strong{Data Frame Columns}
  \enumerate{
     \item \emph{rank}: integer
     \item \emph{teamNumber}: integer
     \item \emph{sortOrder1, sortOrder2, sortOrder3, sortOrder4, sortOrder5,
       sortOrder6}: integer or numeric.
     \item \emph{wins, losses, ties}: integer
     \item \emph{qualAverage}: numeric
     \item \emph{dq}: integer
     \item \emph{matchesPlayed}: integer}

  \strong{Data Frame Attributes}
    \itemize{
    \item \emph{url}: URL submitted to FIRST API
    \item \emph{time_downloaded}: Local System time that the object was
    downladed from the FIRST API server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.}
}
\description{
The results vary depending on the season requested. The 2016 data fields are
listed here. See the FIRST API documentation for data fields for prior
seasons.
}
\details{
See the \emph{Event Rankings} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.

The URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/rankings/event?teamNumber=team&top=top}
}
\examples{
sn <- GetSession("username", "key")
GetRankings(sn, "WAAMV")
GetRankings(sn, "PNCMP", team = 1983)
GetRankings(sn, "ARCHIMEDES", top = 5)
}

