% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetRankings}
\alias{GetRankings}
\title{Get team rankings}
\usage{
GetRankings(session, event, team = NULL, top = NULL)
}
\arguments{
\item{session}{Session A session list created with \code{GetSession()}.}

\item{event}{Character, Case insensitive event code (see \code{GetEvents()}).}

\item{team}{Team number. Optional.}

\item{top}{Integer Number of teams to return.}
}
\value{
A data.frame, json list, or xml list.
   data.frame column names and classes (2016):
     rank: integer
     teamNumber: integer
     sortOrder1, sortOrder2, sortOrder3, sortOrder4, sortOrder5,
       sortOrder6: integer or numeric.
     wins. losses, ties: integer
     qualAverage: numeric
     dq: integer
     matchesPlayed: integer
   Attribute "FIRST_type": "Rankings"
   Attribute "url": URL submitted to FIRST API
}
\description{
The results vary depending on the season requested. The 2016 data fields are
listed here. See the FIRST API documentation for data fields for prior
seasons.
}
\details{
See the \emph{Event Rankings} section of the FIRST API documentation. The
URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/rankings/event?
teamNumber=team&top=top}
}
\examples{
sn <- GetSession(username, key)
GetRankings(sn, 'WAAMV')
GetRankings(sn, "PNCMP", team = 1983)
GetRankings(sn, "ARCHIMEDES", top = 5)
}

