% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapi_http.R
\name{GetSchedule}
\alias{GetSchedule}
\title{Get the match schedule for a specific event.}
\usage{
GetSchedule(session, event, level = "qual", team = NULL, start = NULL,
  end = NULL, expand_cols = FALSE)
}
\arguments{
\item{session}{Session A session list created with \code{GetSession()}.}

\item{event}{Character The FIRST API event code.}

\item{level}{Character Either \code{qual} or \code{playoff}. Defaults to
\code{qual}.}

\item{team}{Integer team number. Optional}

\item{start}{Integer Earliest match to reuturn}

\item{end}{Integer Latest match to return}

\item{expand_cols}{A logical value. Optional, defaults to FALSE. If TRUE, the
dataframe will include one row for each scheduled match, with a different
column for each team. If FALSE, there will be six rows for each match, with
each row listing one assigned team and their station.}
}
\value{
A data.frame, json list, or xml list.
   data.frame column names and classes:
     description: character
     field: character
     tournamentLevel: factor
     matchNumber: integer
     startTime: character

     If expand.rows == FALSE
       Red1.team, Red2.team, Red3.team: factor
       Blue1.team, Blue2.team, Blue3.team: factor
       Red1.surrogate, Red2.surrogate, Red3.surrogate: logical
       Blue1.surrogate, Blue2.surrogate, Blue3.surrogate: logical
     If expand.rows == TRUE
       teamNumber: factor
       alliance: factor (Red, Blue)
       station: factor (Red1, Red2, Red3, Blue1, Blue2, Blue3)
       surrogate: logical
     Attribute "FIRST_type": "Schedule"
     Attribute "url": URL submitted to FIRST API
}
\description{
See the \emph{Event Schedule} section of the FIRST API documentation. The
URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season/schedule/event?
tournamentLevel=level&teamNumber=team&start=start&end=end}
}
\examples{
sn <- GetSession(username, key)
GetSchedule(sn, 'PNCMP')
GetSchedule(sn, 'PNCMP', start=5, end=10)
GetSchedule(sn, 'WAAMV', level='playoff')
GetSchedule(sn, 'PNCMP', team=4911, end=25)
}

