% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetSchedule}
\alias{GetSchedule}
\title{Get the match schedule for a specific event}
\usage{
GetSchedule(session, event, level = "qual", team = NULL, start = NULL,
  end = NULL, mod_since = NULL, only_mod_since = NULL)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}

\item{level}{A character vector containing either \emph{"qual"} or
\emph{"playoff"}. Defaults to \emph{"qual"}. Optional.}

\item{team}{An integer vector containing a team number. Optional.}

\item{start}{An integer vector containing the earliest match to return.
Optional.}

\item{end}{An integer vector containing the latest match to return. Optional.}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to c("data.frame,
  "Schedule"). Returns a logical vector of length one with value \code{NA} if
  data is unchanged since date and time passed in arguments \code{mod_since}
  or \code{only_mod_since}.
}
\description{
Returns either the qualification schedule or the playoff schedule, based on
the value of the \code{level} argument. The \code{start} and \code{end}
arguments allow filtering of results to specific matches.
}
\details{
The data frame returned by \code{GetSchedule()} is in team shape, i.e., each
row contains data for a single team and there are six rows per match. Use
\code{ToAllianceShape()} or \code{ToMatchShape} to convert the data frame to
a three-teams-per-row shape or a six-teams-per-row shape.

See the \emph{Event Schedule} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/schedule/event?
tournamentLevel=level&teamNumber=team&start=start&end=end}
}
\section{Columns}{

  \enumerate{
    \item \emph{description}: character
    \item \emph{field}: character
    \item \emph{level}: factor
    \item \emph{match}: integer
    \item \emph{start}: character
    \item \emph{team}: factor
    \item \emph{alliance}: factor (Red, Blue)
    \item \emph{station}: factor (red1, red2, red3, blue1, blue2, blue3)
    \item \emph{surrogate}: logical}
}

\section{Attributes}{

 \enumerate{
    \item \emph{url}: Character vector containing URL submitted to FIRST API
      server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.
    \item \emph{local_url}: Character vector containing URL used to download
      local data.
    \item \emph{shape}: Character vector that specifies the row and column
      arrangement of the data frame. Initially set to \emph{team},
      indicating there is one team per row and six rows per match. The
      functions \code{ToAllianceShape()} and \code{ToMatchShape} will reshape
      the data frame to have either three teams per row or six teams per row
      and will set the \emph{shape} attribute to \emph{alliance} or
      \emph{match}.
    \item \emph{time_downloaded}: Character vector containing the local
      system time that the object was downladed from the FIRST API server.
      Formatted an an http date and time string.
    \item \emph{last_modified}: Character vector containing the date and time
      that the data was last modified by the FIRST API server.
    \item \emph{mod_since}: Character vector containing the value passed to
      the \code{mod_since argument}, or NULL if no argument was passed.
      Formatted as an http date and time string.
    \item \emph{only_mod_since}: Character vector containing the value passed
      to the \code{only_mod_since} argument, or NULL if no argument was
      passed. Formatted as an http date and time string.}
}
\examples{
sn <- GetSession("username", "key")
GetSchedule(sn, "PNCMP")
GetSchedule(sn, "PNCMP", start=5, end=10)
GetSchedule(sn, "WAAMV", level='playoff')
GetSchedule(sn, "PNCMP", team=4911, end=25)
}

