% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetScores}
\alias{GetScores}
\title{Get detailed match scores.}
\usage{
GetScores(session, event, level = "qual", team = NULL, match = NULL,
  start = NULL, end = NULL)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{event}{A character vector containing a FIRST API event code
(see \code{GetEvents}).}

\item{level}{A character vector containing either \emph{"qual"} or
\emph{"playoff"}. Defaults to \emph{"qual"}. Optional.}

\item{team}{An integer vector containing a team number. Optional. Cannot
specify \code{match} when \code{team} is specified.}

\item{match}{An integer vector containing a match number. Optional. If
specified, \code{GetMatchResults} returns results for only the specified
match. If \code{level} is not specified, returns the results for the
qualification match. To get playoff match results, set \code{level} to
\emph{"playoff"}. Cannot specify \code{team} when \code{match} is
specified.}

\item{start}{An integer vector containing the earliest match to
Optional.}

\item{end}{An integer vector containing the latest match to return. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to
  c("data.frame", "Schedule").

  \strong{Data Frame Columns}
  \enumerate{
     \item \emph{matchLevel}: character
     \item \emph{matchNumber}: integer
     \item \emph{audienceGroup}: character
     \item \emph{alliance}: character
     \item \emph{robot1Auto, robot2Auto, robot3Auto}: character
     \item \emph{autoBouldersLow, autoBouldersHigh}: integer
     \item \emph{teleopBouldersLow, teleopBouldersHigh}: integer
     \item \emph{towerFaceA, towerFaceB, towerFaceC}: character
     \item \emph{towerEndStrength}: integer
     \item \emph{teleopTowerCaptured, teleopDefensesBreached}: logical
     \item \emph{position1, position2, position3, position4,
       position5}: character
     \item \emph{position1Crossings, position2Crossings, position3Crossings,
       position4Crossings, position5Crossings}: integer
     \item \emph{foulCount, techFoulCount}: integer
     \item \emph{autoPoints, autoReachPoints, autoCrossingPoints,
       autoBoulderPoints}: integer
     \item \emph{teleopPoints, teleopCrossingPoints, teleopBoulderPoints,
       teleopChallengePoints, teleopScalePoints}: integer
     \item \emph{breachPoints, capturePoints}: integer
     \item \emph{adustPoints, foulPoints, totalPoints}: integer}

  \strong{Data Frame Attributes}
    \itemize{
    \item \emph{url}: URL submitted to FIRST API
    \item \emph{time_downloaded}: Local System time that the object was
    downladed from the FIRST API server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.}
}
\description{
The results vary depending on the season requested. The 2016 data fields are
listed here. See the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for data fields for prior seasons.
The data frame contains two rows for each match, one for blue and the other
for red.
}
\details{
See the \emph{Detailed Scores} section of the FIRST API documentation for
more details.

The URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season/matches/event/level?
teamNumber=team&matchNumber=match&start=start&end=end}
}
\examples{
sn <- GetSession("username", "key")
GetScores(sn, event = "ARCHIMEDES")
GetScores(sn, event = "WAELL", start = 1, end = 10)
GetScores(sn, event = "WAELL", match = 15)
}

