% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_main.R
\name{GetSeason}
\alias{GetSeason}
\title{Get high-level information for an FRC season}
\usage{
GetSeason(session, mod_since = NULL, only_mod_since = NULL)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}

\item{mod_since}{A character vector containing an HTTP formatted date and
time. Returns \code{NA} if no changes have been made to the requested data
since the date and time provided. Optional.}

\item{only_mod_since}{A character vector containing an HTTP formatted date
and time. This function only returns data that has changed since the date
and time provided. Optional.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to
  c("data.frame, "Season"). Returns a logical vector of length one with value
  \code{NA} if data is unchanged since date and time passed in arguments
  \code{mod_since} or \code{only_mod_since}.
}
\description{
Returns information for the season specified in the session list (see
documentation for the GetSession function for additional details.)
}
\details{
See the \emph{Season Summary} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season}
}
\section{Columns}{

  \enumerate{
    \item \emph{eventCount}: integer
    \item \emph{gameName}: factor
    \item \emph{kickoff}: factor
    \item \emph{rookieStart}: integer
    \item \emph{teamCount}: integer
    \item \emph{FRCChampionships.name}: character
    \item \emph{FRCChampionships.startDate}: character
    \item \emph{FRCChampionships.location}: character}
}

\section{Attributes}{

  \itemize{
    \item \emph{url}: Character vector containing URL submitted to FIRST API
      server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.
    \item \emph{local_url}: Character vector containing URL used to download
      local data.
    \item \emph{time_downloaded}: Character vector containing the local
      system time that the object was downladed from the FIRST API server.
      Formatted an an http date and time string.
    \item \emph{last_modified}: Character vector containing the date and time
      that the data was last modified by the FIRST API server.
    \item \emph{mod_since}: Character vector containing the value passed to
      the \code{mod_since argument}, or NULL if no argument was passed.
      Formatted as an http date and time string.
    \item \emph{only_mod_since}: Character vector containing the value passed
      to the \code{only_mod_since} argument, or NULL if no argument was
      passed. Formatted as an http date and time string.}
}
\examples{
sn <- GetSession("username", "key", season = 2015, staging = TRUE)
summary <- GetSeason(sn)
}

