% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetSeason}
\alias{GetSeason}
\title{Get high-level information for an FRC season}
\usage{
GetSeason(session)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to
  c("data.frame, "Season").

  \strong{Data Frame Columns}
  \enumerate{
    \item \emph{eventCount}: integer
    \item \emph{gameName}: factor
    \item \emph{kickoff}: factor
    \item \emph{rookieStart}: integer
    \item \emph{teamCount}: integer
    \item \emph{FRCChampionships.name}: character
    \item \emph{FRCChampionships.startDate}: character
    \item \emph{FRCChampionships.location}: character}

  \strong{Data Frame Attributes}
  \itemize{
    \item \emph{url}: URL submitted to FIRST API
    \item \emph{time_downloaded}: Local System time that the object was
    downladed from the FIRST API server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.}
}
\description{
Returns information for the season specified in the session list (see
documentation for the GetSession function for additional details.)
}
\details{
See the \emph{Season Summary} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.

The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0/season}
}
\examples{
sn <- GetSession("username", "key", season = 2015, staging = TRUE)
summary <- GetSeason(sn)
}

