% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetSeason}
\alias{GetSeason}
\title{Get high-level information on a particular FRC season.}
\usage{
GetSeason(session)
}
\arguments{
\item{session}{List A list created with \code{GetSession()}.}
}
\value{
JSON or XML formatted text, or an R data frame.
   data.frame column names and classes:
     eventCount: integer
     gameName: factor
     kickoff: factor
     rookieStart: integer
     teamCount: integer
     FRCChampionships.name: character
     FRCChampionships.startDate: character
     FRCChampionships.location: character
   Attribute "FIRST_type": "Season"
   Attribute "url": URL submitted to FIRST API
}
\description{
Returns information for the season specified in the session list (see
documentation for the GetSession function for additional details.)
}
\details{
See the \emph{Season Summary} section of the FIRST API documentation for
additional details. The URL format is:
\code{https://frc-api.firstinspires.org/v2.0/season}
}
\examples{
# staging arg set to TRUE so examples will run without username and key
#   assigned by FIRST.
sn <- GetSession("username", "key", season = 2015, staging = TRUE)
summary <- GetSeason(sn)
}

