% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_http.R
\name{GetServerStatus}
\alias{GetServerStatus}
\title{Get the status of the FIRST API server.}
\usage{
GetServerStatus(session)
}
\arguments{
\item{session}{A Session object created with \code{GetSession()}.}
}
\value{
Depending on the \code{session$format} value, returns JSON text, an
  XML::XMLDocument object, or a data.frame with class set to
  c("data.frame, "Season").

  \strong{Data Frame Columns}
  \enumerate{
    \item \emph{name}: character
    \item \emph{version}: double
    \item \emph{status}: character}

  \strong{Data Frame Attributes}
  \itemize{
    \item \emph{url}: URL submitted to FIRST API
    \item \emph{time_downloaded}: Local System time that the object was
    downladed from the FIRST API server.
    \item \emph{local_test_data}: \code{TRUE} if data was extracted from
      R/sysdata.rda file.}
}
\description{
See the \emph{API Index} section of the FIRST API documentation at
\url{http://docs.frcevents2.apiary.io/#} for more details.
}
\details{
The FIRST API URL format is:

\code{https://frc-api.firstinspires.org/v2.0}
}
\examples{
sn <- GetSession("username", "key")
status <- GetServerStatus(sn)
}

