% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstapiR_util.R
\name{MergeResults}
\alias{MergeResults}
\title{Merge MatchResults and Scores data frames into a single data frame.}
\usage{
MergeResults(matchresults.df, scores.df)
}
\arguments{
\item{matchresults.df}{A MatchResults class data frame.}

\item{scores.df}{A Scores class data frame.}
}
\value{
A Results class data frame containing both the teams assigned to a
  match and their detailed scores.
}
\description{
Merge MatchResults and Scores data frames into a single data frame.
}

